Move collection exercise steps to their own file'
